+++ core/pv-tango/src/main/java/org/phoebus/pv/tango/TangoDeviceHelper.java
@@ -300,7 +300,7 @@ public class TangoDeviceHelper {
             timeOut = DEFAULT_TIMEOUT;
         } else {
             Integer tmp = TIME_OUT_MAP.get(device);
-            timeOut = tmp == null ? DEFAULT_TIMEOUT : tmp.intValue();
+            timeOut = tmp == null ? DEFAULT_TIMEOUT : tmp;
         }
         return timeOut;
     }
@@ -317,7 +317,7 @@ public class TangoDeviceHelper {
             if (timeOut < 1) {
                 timeOut = DEFAULT_TIMEOUT;
             }
-            TIME_OUT_MAP.put(device, Integer.valueOf(timeOut));
+            TIME_OUT_MAP.put(device, timeOut);
             getDeviceProxy(device, false);
         }
     }
